requirements.txt
ics
mistune
python-slugify
